frameworks/python/django/djangowallet/requirements.txt
# main
Django==2.0
django-admin-env-notice==0.1.3
psycopg2-binary==2.8.1
psycopg2==2.7
djangorestframework==3.9.2
coreapi==2.3.3
django-migrations-graph==0.1.1
django-eraserhead==0.1.4
django-debug-toolbar==1.9
djangorestframework-csv==2.1.0

# Tests / linters
pycodestyle==2.4
pyflakes==1.5.0
mccabe==0.5.2
pylama==7.7.1
